requirements.txt
anthropic>=0.39.0
openai>=1.54.0
tavily-python>=0.5.0
pydantic>=2.9.0
python-dotenv>=1.0.1
yfinance>=0.2.40
rich>=13.9.0
